[BUGFIX] Workaround for inline FAL field
Inline FAL fields does not support drag and drop, if they are inside of a Flexform. This seems to be a core bug. To avoid the circumstance not being able to change sorting of media (without deleting items) the sample configuration template for inline FAL has disabled dragdrop and enabled sort instead.

Additionally the use of FAL in group fields is not recommended (as I've learned on T3DD14). The snippet got a notice for that, now.

Resolves: #58987
Change-Id: I7ede1e4feed2a585178199f3608eeb1bff26657f
Reviewed-on: https://review.typo3.org/31092
Reviewed-by: Armin Ruediger Vieweg
Tested-by: Armin Ruediger Vieweg