[TASK] Change file template fields to group
The wizard of TYPO3 is able to select files, but adds a FAL string to template
like "file:123". Editors which have no access to this file get errors when saving
content elements based on this DCE.

The new group field (using file_reference) stored the path to template to database.

Resolves: #63079
Change-Id: Ic9bc06f8008d4898c3e060468d9df9df7c57da53
Reviewed-on: http://review.typo3.org/34732
Tested-by: Philipp Thiele <philipp.thiele@phth.de>
Reviewed-by: Armin Ruediger Vieweg <armin@v.ieweg.de>
Tested-by: Armin Ruediger Vieweg <armin@v.ieweg.de>