[TASK] Ensure gridelements compatibility
When you drag new content elements to a grid or adding a ce in a grid column, may cause errors if the field "tx_gridelements_container" is missing in DCE palette fields.

With this patch, the two fields "colPos" and "tx_gridelements_container" will be added by DCE cache to the end of palette, if gridelements is installed and these fields are missing.