[FEATURE] Change flexform syntax for tabs (sheets)
In the past DCE named tabs in flexform configuration like this:
<sheet0></sheet0>

But this has the effect that all your data is broken, when you change
the order of tabs in a DCE. Now the sheets have a named identifier. You
can set the identifier in the variable field which is also visible for
tab fields, now.

The flexform configuration looks like this now:
<sheet.tabGeneral></sheet.tabGeneral>

The very first sheet has the identifier/variable "tabGeneral" by default.