[TASK] Add check to DCE's cache generator if database tables are existing
Because the extension "typo3_console" triggers DCEs cache generator, when
typo3_console should update the database schema. But at this point no tx_dce_
tables are existing (because we are in the step to create them).

This causes exceptions. It still works, but to avoid these exceptions the cache
generator checks first if the required tables are existing, and if not it returns
an empty array, instead of an array with found DCEs.