[FEATURE] Add update script to fix malformed variable names in DceFields
In older versions of DCE the userfunction, which checks and corrects entered
variable names to be owerCamelCase, did not work properly. But when editing
such old DCEs the variables became corrected and does not match with
flexform structure used in tt_content.

This update checks if such fields exist and correct them in tt_content's
column and in DceFields.

It does not correct the fluid templates for you!

Resolves: #77457