[CLEANUP] Reduce code duplication and avoid internal sub requests
Instead of creating an internal sub request to the "renderDce" action
of the DceController move the required code for instanciating a DCE
domain model object to an existing utility class.

This avoids expensive bootstraping and calling of a controller action
and additionally avoids the serialization/unserialization of whole
domain objects.