[BUGFIX] do not throw and FATAL when creating new DCE records
When creating a new, default content element record and then switching the content element type to be a DCE
(not using the "create new content element"-wizard; e.g. when creating tt_content inline records in for example a news record)
the record is initially saved, but pi_flexform happens to be empty at this point.
This causes the second argument of DceRepository::processFillingFields to be null instead of array, which throws an PHP FATAL.