[FEATURE] Allow partials and layouts fallback
DCE will first check for partials and layouts in the folder defined
in the DCE database record. And then, if not found or empty, in the
default folder of the DCE extension.

This allows developers to put just custom partials/layouts in the site
package. Before they were forced to copy all default partials as well.