[BUGFIX] Do not use Extbase Controller Action to return DCE Instance
When a DCE field contains FAL resources, an exception occured, that serialization of Closures are not allowed.

This happened, when DceContainer builds instances of DCEs. To build an instance, DCE instantiated Extbase a second time and executed the DceController::renderDceAction.
Because Extbase does not allow to return objects in actions, I've serialized (and gzipped) the model.

With this commit this part is totally refactored:
- No un/serialization anymore
- No second bootstrapping of Extbase (utility removed)
- Moved database related stuff from DceController to DceRepository