[BUGFIX] Fix TtContentLabel userfunction in case of incomplete data
In some cases, incomplete data may be passed to
TtContentLabel::getLabel(). Due to this we perform a check for missing
field 'CType' first.

This fixes a problem if editing of the pi_flexform field of dce
content elements is invoked by using the Fluid ViewHelper link.editRecord
with parameter "fields", for example, e.g.

<be:link.editRecord uid="{uid}" table="tt_content"
  fields="pi_flexform">edit record {uid} pi_flexform
</be:link.editRecord>

Resolves: #102